refactor: decouple optimization metrics and logging logic into dedicated header and source files
