+++ src/optimization_logger.cpp
@@ -1,0 +1,3 @@
+#include "optimization_logger.h"
+
+// Dedicated compilation unit for optimization logger
